+++ IconicFramework/Constants.cs
@@ -8,5 +8,5 @@ internal static class Constants
 
     public const string ModId = "furyx639.ToolbarIcons";
 
-    public const string UiPath = "/UI";
+    public const string UiPath = ModId + "/UI";
 }