+++ IconicFramework/Integrations/ToDew.cs
@@ -1,10 +1,10 @@
-namespace LeFauxMods.IconicFramework.Integrations;
-
 using System.Reflection;
-using Common.Integrations.IconicFramework;
+using LeFauxMods.Common.Integrations.IconicFramework;
+using LeFauxMods.IconicFramework.Utilities;
 using Microsoft.Xna.Framework;
 using StardewValley.Menus;
-using Utilities;
+
+namespace LeFauxMods.IconicFramework.Integrations;
 
 /// <summary>Mod integration with ToDew.</summary>
 internal sealed class ToDew